Everything except subject literal further filtering
