+++ src/components/SearchFilters/SearchFilters.tsx
@@ -48,7 +48,7 @@ const SearchFilters = ({ aggregations }: { aggregations?: Aggregation[] }) => {
     }
     router.push(
       {
-        pathname: "/search",
+        pathname: router.pathname,
         query: updatedQuery,
       },
       undefined,
